+++ src/main/java/muzusi/config/security/SecurityFilterConfig.java
@@ -1,0 +1,29 @@
+package muzusi.config.security;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import lombok.RequiredArgsConstructor;
+import muzusi.global.security.filter.JwtAuthenticationFilter;
+import muzusi.global.security.filter.JwtExceptionFilter;
+import muzusi.global.security.jwt.JwtProvider;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.core.userdetails.UserDetailsService;
+
+@Configuration
+@RequiredArgsConstructor
+public class SecurityFilterConfig {
+
+    private final UserDetailsService userDetailsService;
+    private final JwtProvider jwtProvider;
+    private final ObjectMapper objectMapper;
+
+    @Bean
+    public JwtAuthenticationFilter jwtAuthenticationFilter() {
+        return new JwtAuthenticationFilter(userDetailsService, jwtProvider, objectMapper);
+    }
+
+    @Bean
+    public JwtExceptionFilter jwtExceptionFilter() {
+        return new JwtExceptionFilter(objectMapper);
+    }
+}
